refactor: remove paths
