+++ src/tsconfig.json
@@ -7,13 +7,7 @@
 		"composite": true,
 		"preserveConstEnums": true,
 		"useDefineForClassFields": false,
-		"noImplicitReturns": true,
-		"paths": {
-			"#structures/*": ["structures/*"],
-			"#client/*": ["client/*"],
-			"#types/*": ["types/*"],
-			"#utils/*": ["utils/*"]
-		}
+		"noImplicitReturns": true
 	},
 	"include": ["."]
 }
